model/requirements.txt
Pillow
boto3
requests
python-dotenv
ultralytics
pascal-voc